[Resource] Combine Postgres database and pool
