+++ config/prod.yaml
@@ -7,7 +7,7 @@ server:
 plugins:
   resources:
     database:
-      type: pipeline.plugins.resources.postgres_database:PostgresDatabaseResource
+      type: pipeline.plugins.resources.postgres:PostgresResource
       host: "${DB_HOST}"
       port: 5432
       name: "${DB_USERNAME}"
@@ -27,7 +27,7 @@ plugins:
     memory:
       type: pipeline.plugins.resources.memory:MemoryResource
       database:
-        type: pipeline.plugins.resources.postgres_database:PostgresDatabaseResource
+        type: pipeline.plugins.resources.postgres:PostgresResource
         host: "${DB_HOST}"
         port: 5432
         name: "${DB_NAME}"
