+++ tests/integration/test_chat_history_backends.py
@@ -6,22 +6,15 @@ from pathlib import Path
 import pytest
 
 from config.environment import load_env
-from pipeline import (
-    ConversationEntry,
-    MetricsCollector,
-    PipelineStage,
-    PipelineState,
-    PluginContext,
-    PluginRegistry,
-    ResourceRegistry,
-    SystemRegistries,
-    ToolRegistry,
-)
+from pipeline import (ConversationEntry, MetricsCollector, PipelineStage,
+                      PipelineState, PluginContext, PluginRegistry,
+                      ResourceRegistry, SystemRegistries, ToolRegistry)
 from pipeline.plugins.prompts.chat_history import ChatHistory
 from pipeline.plugins.resources.duckdb_database import DuckDBDatabaseResource
-from pipeline.plugins.resources.in_memory_storage import InMemoryStorageResource
+from pipeline.plugins.resources.in_memory_storage import \
+    InMemoryStorageResource
 from pipeline.plugins.resources.memory import MemoryResource
-from pipeline.plugins.resources.postgres_database import PostgresDatabaseResource
+from pipeline.plugins.resources.postgres import PostgresResource
 from pipeline.plugins.resources.sqlite_storage import SQLiteStorageResource
 
 load_env(Path(__file__).resolve().parents[2] / ".env")
@@ -29,13 +22,14 @@ load_env(Path(__file__).resolve().parents[2] / ".env")
 
 async def run_history_test(resource):
     await getattr(resource, "initialize", lambda: None)()
-    if isinstance(resource, PostgresDatabaseResource):
-        await resource._connection.execute("DROP TABLE IF EXISTS test_history")
-        await resource._connection.execute(
-            "CREATE TABLE test_history ("
-            "conversation_id text, role text, content text, "
-            "metadata jsonb, timestamp timestamptz)"
-        )
+    if isinstance(resource, PostgresResource):
+        async with resource.connection() as conn:
+            await conn.execute("DROP TABLE IF EXISTS test_history")
+            await conn.execute(
+                "CREATE TABLE test_history ("
+                "conversation_id text, role text, content text, "
+                "metadata jsonb, timestamp timestamptz)"
+            )
     if isinstance(resource, SQLiteStorageResource):
         resource._table = "test_history"
         await resource.initialize()
@@ -106,7 +100,7 @@ def test_postgres_history():
         "history_table": "test_history",
     }
     try:
-        resource = PostgresDatabaseResource(cfg)
+        resource = PostgresResource(cfg)
         asyncio.run(resource.initialize())
     except OSError as exc:
         pytest.skip(f"PostgreSQL not available: {exc}")
