+++ tests/integration/test_postgres_history.py
@@ -6,7 +6,7 @@ from pathlib import Path
 import pytest
 
 from config.environment import load_env
-from pipeline.plugins.resources.postgres_database import PostgresDatabaseResource
+from pipeline.plugins.resources.postgres import PostgresResource
 from pipeline.state import ConversationEntry
 
 load_env(Path(__file__).resolve().parents[2] / ".env")
@@ -26,23 +26,24 @@ CONN = {
 @pytest.mark.integration
 def test_save_and_load_history():
     async def run():
-        resource = PostgresDatabaseResource(CONN)
+        resource = PostgresResource(CONN)
         try:
             await resource.initialize()
         except OSError as exc:
             pytest.skip(f"PostgreSQL not available: {exc}")
-        await resource._connection.execute("DROP TABLE IF EXISTS test_history")
-        await resource._connection.execute(
-            "CREATE TABLE test_history ("
-            "conversation_id text, role text, content text, "
-            "metadata jsonb, timestamp timestamptz)"
-        )
+        async with resource.connection() as conn:
+            await conn.execute("DROP TABLE IF EXISTS test_history")
+            await conn.execute(
+                "CREATE TABLE test_history ("
+                "conversation_id text, role text, content text, "
+                "metadata jsonb, timestamp timestamptz)"
+            )
         entries = [
             ConversationEntry(content="hello", role="user", timestamp=datetime.now())
         ]
         await resource.save_history("conv1", entries)
         loaded = await resource.load_history("conv1")
-        await resource._connection.close()
+        await resource._pool.close()
         return loaded
 
     history = asyncio.run(run())
