+++ tests/test_postgres_resource.py
@@ -6,7 +6,7 @@ from unittest.mock import AsyncMock, patch
 import asyncpg
 
 from config.environment import load_env
-from pipeline.plugins.resources.postgres_database import PostgresDatabaseResource
+from pipeline.plugins.resources.postgres import PostgresResource
 
 load_env(Path(__file__).resolve().parents[1] / ".env")
 
@@ -21,30 +21,38 @@ async def init_resource():
     }
     conn = AsyncMock()
     conn.fetchval = AsyncMock(return_value=1)
-    with patch("asyncpg.connect", new=AsyncMock(return_value=conn)) as mock_connect:
-        plugin = PostgresDatabaseResource(cfg)
+    pool = AsyncMock()
+    pool.acquire = AsyncMock(return_value=conn)
+    pool.release = AsyncMock()
+    with patch("asyncpg.create_pool", new=AsyncMock(return_value=pool)) as mock_pool:
+        plugin = PostgresResource(cfg)
         await plugin.initialize()
-        mock_connect.assert_awaited_with(
+        mock_pool.assert_awaited_with(
             database="db",
             host=os.environ["DB_HOST"],
             port=5432,
             user=os.environ["DB_USERNAME"],
             password=os.environ.get("DB_PASSWORD", ""),
+            min_size=1,
+            max_size=5,
         )
     return plugin, conn
 
 
 def test_initialize_opens_connection():
     plugin, _ = asyncio.run(init_resource())
-    assert plugin._connection is not None
+    assert plugin._pool is not None
     assert plugin.has_vector_store
 
 
 async def run_health_check():
-    plugin = PostgresDatabaseResource({})
+    plugin = PostgresResource({})
     conn = AsyncMock()
     conn.fetchval = AsyncMock(return_value=1)
-    plugin._connection = conn
+    pool = AsyncMock()
+    pool.acquire = AsyncMock(return_value=conn)
+    pool.release = AsyncMock()
+    plugin._pool = pool
     result = await plugin.health_check()
     conn.fetchval.assert_awaited_with("SELECT 1")
     return result
@@ -66,8 +74,11 @@ def test_malicious_table_name_is_quoted():
             "history_table": "history; DROP TABLE y;",
         }
         conn = AsyncMock()
-        with patch("asyncpg.connect", new=AsyncMock(return_value=conn)):
-            plugin = PostgresDatabaseResource(cfg)
+        pool = AsyncMock()
+        pool.acquire = AsyncMock(return_value=conn)
+        pool.release = AsyncMock()
+        with patch("asyncpg.create_pool", new=AsyncMock(return_value=pool)):
+            plugin = PostgresResource(cfg)
             await plugin.initialize()
             expected_table = (
                 f'{asyncpg.utils._quote_ident(cfg["db_schema"])}.'
